*: Remove serialized timeline generation
Remove the generation of a serialized timeline and related timeline "aggregate"
that is one of the reasons why we currently have all the commands being serialized.

Instead, detect every new timeline using the event groupID, every group of
events triggered by a commands will have a different groupID, we are not using
the correlationID since multiple commands could be correlated and have the same
correlationID, we are not using the causationID because multiple commands may be
triggered by the same event and so have the same causationID. The timeline
timestamp is the timestamp of the first event in that groupID.

The TimeLineNumber that is replacing the TimeLineSequenceNumber is dinamically
generated from the timeline timestamp converted to unixnano.