*: improve error handling
Try to provide a proper stack trace without adding redundant stack traces. For
how pkg/errors works this is currently a manual and error prone work. The rules
are these:

* Use errors.New/Errorf when creating a new error, this will save a stacktrace
* Use errors.Wrap/Wrapf/WithStack when calling an external library that doens't return an
error with a stacktrace
* Just return the errors or use errors.WithMessage when returning an error
generated inside the app or a library that returns an error that already have
a stacktrace
* For errors returned with a channel by another go routine it's useful to record
also another stacktrace using errors.Wrap/Wrapf/WithStack.

Also use #+v when logging errors to also print the stacktrace (we aren't
currently using zap as a structured logger, in that case zap will automatically
add an ErrorVerbose field on errors with a stacktrace).