eventstore: initial concurrent writes implementation
The eventstore now accepts concurrent writes. They are currently masked by the
serializable isolation level due to the current architecture of a single tx for
every command shared also with the readdb.

This will be relaxed in future when moving to a completely decoupled eventstore
and readdb.

The WriteEvents method requires that all the events are for the same aggregate
and will check that the provided current aggregate version is the current one in
the store to check concurrent writes to the same aggregate.

Since in future there can be many concurrent transactions the above check won't
work with concurrent transactions but it's enforced by an unique constraint on
the aggregate event version that will make the transaction fail.

The global sequence number is now provided by a sequence so it could contain
gaps.

To achieve ordered sequence numbers commits (this won't be possible by default due to
concurrent transactions that could commit sequence numbers out of order) an
advisory lock is taken (for the smaller possible time) before inserting the
events.