saga: add member request saga
This patch implements the member request saga.

When a member creation or update is requested it's saved in a new aggregate
called memberchange. Its events are handled by the memberrequesthandler that
will call a saga for every memberchange. The saga checks that the username,
email and matchUID aren't already reserved calling a reserve command on an
uniquevalueregistry aggregate that is responsible for an single username or
email or matchUID.

If one of the reservations fails the previous one are rolled back and the saga
ends.

If the memberchange is an update the old reserved values are released.

This is a CQRS/ES way to handle unique constraints in a distributed, scalable
way using a long running transaction (saga).

This is implemented considering that every command executed on an aggregate and
related events can fails to commit to the eventstore for multiple reasons. So
the saga can break in multiple states. To handle this all the aggregates command
must be idempotent so they can be executed also more than one time without
causing problems.