*: implement a real cqrs+es
Use the aggregates, eventhandlers and saga implemented in previous patches.

The eventstore and the readdb are completely decoupled.
The readdb is an eventhandler that will listen on new events and apply them.

Every write to the eventstore is a single transaction.

Change the postgres default transaction isolation level from serializable to
repeatable read since we are able to achieve more concurrency maintaing
consistency without the protection of the serializable level.

The command services, after data validation, executes a command on the related
aggregate. Every command services method operates on only one aggregates.

To keep synchronous client response, the graphql api will wait for the generated
events to be applied in the readdb.

If a command (like createMember) will trigger a saga, we'll wait for the saga to
be completed.

More docs/blogs will come to explain all the choices and how they works.